Modified Indicators_statistics.csv after rerunning first few cells of Figures_SI.ipynb
Not committing the notebook itself, since executing further cells gives
errors. Just committing the updated CSV file, to be able to inspect what
has changed. Some numbers have changed a lot. Is this an error, or have
there been significant changes to the data in the IIASA database since
the original paper was published (we use the "ecemf" open database,
rather than the "ecemf_internal" database referenced by the original
code, which is not open).